refactor: Profits 객체명을 BettingRate로 변경
