+++ src/main/java/blackjack/model/state/Finished/Bust.java
@@ -1,9 +1,9 @@
 package blackjack.model.state.Finished;
 
-import static blackjack.model.Profits.DRAW;
-import static blackjack.model.Profits.LOSE;
+import static blackjack.model.BettingRate.DRAW;
+import static blackjack.model.BettingRate.LOSE;
 
-import blackjack.model.Profits;
+import blackjack.model.BettingRate;
 import blackjack.model.card.Cards;
 import blackjack.model.state.State;
 
@@ -24,7 +24,7 @@ public class Bust extends Finished {
     }
 
     @Override
-    public Profits calculateProfit(State otherState) {
+    public BettingRate calculateBettingRate(State otherState) {
         if (otherState.isBust()) {
             return DRAW;
         }
